+++ src/test/kotlin/de/xeroli/kbool/BoolTest.kt
@@ -8,11 +8,22 @@ import kotlin.test.assertTrue
 class BoolTest {
     @Test
     fun testSupplier() {
-        var a : String? = null
-        var notNull = (a != null).asBool("String is not null")
-        var longEnough = Bool.SupplierBool( {(a!!.length > 7).asBool("String has at least 7 characters")})
-        println(notNull and longEnough)
-        println((notNull and longEnough).getCause())
+        var a: String? = null
+        var notNull = Bool.of{a != null}.named("String is not null")
+        var longEnough = Bool.of{ (a!!.length > 7) }.named("String has at least 7 characters")
+
+        assertFalse((notNull and longEnough).isTrue(), "false because of a is null")
+        assertEquals("String is not null - false", (notNull and longEnough).getCause(), "should be 'String is not null - false'")
+
+        a = "Hallo Welt!"
+        assertTrue((notNull and longEnough).isTrue(), "false because of a is null")
+        assertEquals("String is not null - true, String has at least 7 characters - true", (notNull and longEnough).getCause(), "should be 'String is not null - false'")
+
+        var i = 1
+        val bool = Bool.of { i > 10 }.named("i greater than 10?")
+        assertFalse(bool.booleanValue())
+        i = 100
+        assertTrue(bool.booleanValue())
     }
 
     @Test
@@ -58,18 +69,18 @@ class BoolTest {
     fun testEquality() {
         var directlyNamed = true.asBool("directly")
         var lateNamed = true.asBool().named("directly")
-        assertFalse( directlyNamed == lateNamed, "false ")
+        assertFalse(directlyNamed == lateNamed, "false ")
 
         directlyNamed.booleanValue()
         lateNamed.booleanValue()
-        assertTrue( directlyNamed == lateNamed, "naming time has no impact on equality")
+        assertTrue(directlyNamed == lateNamed, "naming time has no impact on equality")
 
         directlyNamed = true.asBool("directly")
         lateNamed = true.asBool().named("directly")
 
         // implicitly evaluate one of them
         lateNamed.booleanValue()
-        assertFalse( directlyNamed == lateNamed, "naming time has no impact on equality")
+        assertFalse(directlyNamed == lateNamed, "naming time has no impact on equality")
 
         var aNot = !directlyNamed
         var otherNot = !directlyNamed
@@ -79,4 +90,5 @@ class BoolTest {
         otherNot = directlyNamed.not().named("nonono")
         assertFalse(aNot == otherNot, "two not's of same bool but renamed should not be equal")
     }
+
 }